+++ src/libs/flows.ts
@@ -1,5 +1,7 @@
-import { WorkFlow, FlowRuntime, LaunchHandler } from './global.type';
-import { isPromise } from './util';
+import {
+  WorkFlow, FlowRuntime, LaunchHandler, BlockFlowConfig, DebounceFlowConfig,
+} from './global.type';
+import { isPromise, noop } from './util';
 
 function promise():WorkFlow {
   return function process(runtime:FlowRuntime):LaunchHandler {
@@ -41,7 +43,9 @@ export class Flows {
     };
   }
 
-  static debounce(ms:number, leading?:boolean):WorkFlow {
+  static debounce(time:number|DebounceFlowConfig, lead?:boolean):WorkFlow {
+    const ms = typeof time === 'number' ? time : time.time;
+    const leading = typeof time === 'number' ? lead : time.leading;
     if (leading) {
       return function leadingProcess(runtime:FlowRuntime):LaunchHandler {
         const { state } = runtime;
@@ -66,6 +70,47 @@ export class Flows {
             state.id = setTimeout(() => {
               method(...args);
             }, ms);
+          };
+        },
+      };
+    };
+  }
+
+  static block(config?:number|BlockFlowConfig):WorkFlow {
+    const timeout = (function computeTimeout() {
+      if (config == null) {
+        return undefined;
+      }
+      if (typeof config === 'number') {
+        return config;
+      }
+      return config.timeout;
+    }());
+    function processPromiseFinished(rt:Promise<unknown>, time?:number):Promise<unknown> {
+      const mainPromise = Promise.resolve(rt).then(noop, noop);
+      const timeoutPromise = time == null
+        ? mainPromise : new Promise<void>((r) => setTimeout(r, time));
+      return Promise.race([mainPromise, timeoutPromise]);
+    }
+    return function process(runtime:FlowRuntime):LaunchHandler {
+      const { state } = runtime;
+      return {
+        invoke(method) {
+          return function blockMethod(...args:any[]) {
+            if (state.running) {
+              return state.result;
+            }
+            state.running = true;
+            const result = method(...args);
+            state.result = result;
+            if (!isPromise(result)) {
+              state.running = false;
+              return state.result;
+            }
+            processPromiseFinished(result, timeout).then(() => {
+              state.running = false;
+            });
+            return state.result;
           };
         },
       };